Removed unused imports and unnecessary casts

+++ src/main/java/net/pms/dlna/DLNAResource.java
@@ -2516,7 +2516,7 @@ public abstract class DLNAResource extends HTTPResource implements Cloneable, Ru
 				if (range.isTimeRange()) {
 					resume.update((Range.Time) range, this);
 				}
-				params.timeseek = (long) (resume.getTimeOffset() / 1000);
+				params.timeseek = resume.getTimeOffset() / 1000;
 				if (player == null) {
 					player = new FFMpegVideo();
 				}
